Integrate Linkoping data (#25)
Adam Staadig and Andread Tillmar (Linköping University) presented a poster evaluating a panel of 40-45 microhaplotypes at ISFG 2019. Adam was kind enough to share the marker data and allele frequencies for a Swedish cohort, and give permission for me to publish them. All of the variants were previously captured due to proximity to markers in ALFRED. Indeed, there is substantial overlap between these new markers and the ALFRED markers.

This update integrates the new marker and frequency data into MicroHapDB, with the source label `ISFG2019:P597`.