Integrate 1000 Genomes Project population definitions (#45)
While preparing for #42, I discovered that the 26 global populations from the 1000 Genomes Project are included in the 96 ALFRED populations, under different IDs. This update adds a new data source to the `dbbuild` directory, and the 26 1KGP population definitions were moved from `dbbuild/sources/alfred/` to `dbbuild/sources/1kgp/`. At the moment, the `marker.tsv` and `frequency.tsv` files in `dbbuild/sources/1kgp` are empty, but frequency data will be added as part of #42.

Closes #44.