Add "no truncate" mode for marker tabular output (#48)
Previously I had tried a couple of different things to try to coerce Pandas dataframe objects to print non-truncated tables. The microhapdb/conftest.py was an attempt to do so in the testing environment. This update adds a CLI flag to microhapdb marker to disable column truncation (it is enabled by default). Column truncation shouldn't be an issue for any other data type, but if it were it would be straightforward to implement a similar approach there.